requirements.txt
altgraph==0.16.1
auto-py-to-exe==2.5.2
bottle==0.12.16
bottle-websocket==0.2.9
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
Eel==0.10.4
Flask==1.0.2
Flask-Sockets==0.2.1
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
macholib==1.11
MarkupSafe==1.1.1
numpy==1.16.1
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pefile==2018.8.8
Pillow==5.4.1
pycparser==2.19
PyInstaller==3.4
pytesseract==0.2.6
pytube==9.5.0
pywin32-ctypes==0.2.0
requests==2.22.0
urllib3==1.25.2
Werkzeug==0.14.1
whichcraft==0.5.2
